Refactor bot into package and add tooling
