+++ .devcontainer/devcontainer.json
@@ -1,7 +1,7 @@
 {
   "name": "discord-lm-app",
   "image": "mcr.microsoft.com/devcontainers/python:3.11",
-  "postCreateCommand": "pip install -r requirements.txt",
+  "postCreateCommand": "pip install -r requirements.txt && black . && ruff --fix .",
   "customizations": {
     "vscode": {
       "extensions": ["ms-python.python"]
