+++ .gitignore
@@ -1,2 +1,4 @@
 .env
 __pycache__/
+.venv/
+.pytest_cache/
