+++ CHANGELOG.md
@@ -1,11 +1,12 @@
 # Changelog
 
-## Unreleased
+All notable changes to this project will be documented in this file.
 
-## v0.1.0 - 2025-05-20
-- Google CSE integration.
-- Search returns eight results.
-- URL-safe 2 000-char splitter.
-- Fixed blank-message bug.
-- Inline citation rule.
+The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/)
+and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## [0.1.0] – 2025-05-20
+### Added
+- Initial bot with OpenAI chat, Google CSE search, URL-safe splitter, mention trigger.
