+++ pyproject.toml
@@ -1,0 +1,6 @@
+[tool.black]
+line-length = 100
+
+[tool.ruff]
+line-length = 100
+fix = true
