+++ src/discord_lm_bot/__init__.py
@@ -1,0 +1,3 @@
+from .discord_bot import run_bot, query_chatgpt
+
+__all__ = ["run_bot", "query_chatgpt"]
